Extract CalendarSyncAccountMixin from GmailAccount
